Automatically register context manager types; clean up some old config code
