+++ Source/Csla/Configuration/Fluent/CslaDataPortalConfiguration.cs
@@ -6,8 +6,8 @@
 // <summary>Use this type to configure the settings for CSLA .NET</summary>
 //-----------------------------------------------------------------------
 using System;
-using System.Collections.Generic;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
 
 namespace Csla.Configuration
 {
@@ -22,6 +22,23 @@ namespace Csla.Configuration
     public static CslaDataPortalConfiguration DataPortal(this ICslaConfiguration config)
     {
       return new CslaDataPortalConfiguration(config);
+    }
+
+    /// <summary>
+    /// Add services required to host the server-side
+    /// data portal.
+    /// </summary>
+    /// <param name="builder"></param>
+    public static CslaDataPortalConfiguration AddServerSideDataPortal(this CslaDataPortalConfiguration builder)
+    {
+      var services = builder.Services;
+      services.TryAddTransient(typeof(Server.IDataPortalServer), typeof(Csla.Server.DataPortal));
+      services.TryAddTransient<Server.DataPortalSelector>();
+      services.TryAddTransient<Server.SimpleDataPortal>();
+      services.TryAddTransient<Server.FactoryDataPortal>();
+      services.TryAddTransient<Server.DataPortalBroker>();
+      services.TryAddSingleton(typeof(Server.Dashboard.IDashboard), typeof(Csla.Server.Dashboard.NullDashboard));
+      return builder;
     }
   }
 
@@ -50,154 +67,132 @@ namespace Csla.Configuration
       CslaConfiguration = config;
     }
 
-    //internal bool UseDataPortalServer { get; set; } = false;
-
     ///// <summary>
-    ///// Register services necessary for the
-    ///// server-side data portal to function.
+    ///// Sets the type of the IDataPortalActivator provider.
     ///// </summary>
-    //public CslaDataPortalConfiguration EnableDataPortalServer()
+    ///// <param name="type">Activator type</param>
+    //public CslaDataPortalConfiguration ActivatorType(Type type)
     //{
-    //  UseDataPortalServer = true;
+    //  ActivatorType(type.AssemblyQualifiedName);
     //  return this;
     //}
 
-    /// <summary>
-    /// Sets the type of the IDataPortalActivator provider.
-    /// </summary>
-    /// <param name="type">Activator type</param>
-    public CslaDataPortalConfiguration ActivatorType(Type type)
-    {
-      ActivatorType(type.AssemblyQualifiedName);
-      return this;
-    }
-
-    /// <summary>
-    /// Sets the type of the IDataPortalActivator provider.
-    /// </summary>
-    /// <param name="typeName">Assembly qualified type name</param>
-    public CslaDataPortalConfiguration ActivatorType(string typeName)
-    {
-      ConfigurationManager.AppSettings["CslaDataPortalActivator"] = typeName;
-      return this;
-    }
-
     ///// <summary>
-    ///// Sets an instance of the IDataPortalActivator provider.
+    ///// Sets the type of the IDataPortalActivator provider.
     ///// </summary>
-    ///// <param name="activator">Activator instance</param>
-    //public CslaDataPortalConfiguration Activator(Server.IDataPortalActivator activator)
+    ///// <param name="typeName">Assembly qualified type name</param>
+    //public CslaDataPortalConfiguration ActivatorType(string typeName)
     //{
-    //  ApplicationContext.DataPortalActivator = activator;
+    //  ConfigurationManager.AppSettings["CslaDataPortalActivator"] = typeName;
     //  return this;
     //}
 
-    /// <summary>
-    /// Sets the authentication type being used by the
-    /// CSLA .NET framework.
-    /// </summary>
-    /// <param name="typeName">Authentication type value (defaults to 'Csla')</param>
-    public CslaDataPortalConfiguration AuthenticationType(string typeName)
-    {
-      ConfigurationManager.AppSettings["CslaAuthentication"] = typeName;
-      return this;
-    }
+    ///// <summary>
+    ///// Sets the authentication type being used by the
+    ///// CSLA .NET framework.
+    ///// </summary>
+    ///// <param name="typeName">Authentication type value (defaults to 'Csla')</param>
+    //public CslaDataPortalConfiguration AuthenticationType(string typeName)
+    //{
+    //  ConfigurationManager.AppSettings["CslaAuthentication"] = typeName;
+    //  return this;
+    //}
 
-    /// <summary>
-    /// Sets the type name to be used for server-side data portal
-    /// authorization. Type must implement IAuthorizeDataPortal.
-    /// </summary>
-    /// <param name="type">Authorization provider type</param>
-    public CslaDataPortalConfiguration ServerAuthorizationProviderType(Type type)
-    {
-      ServerAuthorizationProviderType(type.AssemblyQualifiedName);
-      return this;
-    }
+    ///// <summary>
+    ///// Sets the type name to be used for server-side data portal
+    ///// authorization. Type must implement IAuthorizeDataPortal.
+    ///// </summary>
+    ///// <param name="type">Authorization provider type</param>
+    //public CslaDataPortalConfiguration ServerAuthorizationProviderType(Type type)
+    //{
+    //  ServerAuthorizationProviderType(type.AssemblyQualifiedName);
+    //  return this;
+    //}
 
-    /// <summary>
-    /// Sets the type name to be used for server-side data portal
-    /// authorization. Type must implement IAuthorizeDataPortal.
-    /// </summary>
-    /// <param name="typeName">Assembly qualified type name</param>
-    public CslaDataPortalConfiguration ServerAuthorizationProviderType(string typeName)
-    {
-      ConfigurationManager.AppSettings["CslaAuthorizationProvider"] = typeName;
-      return this;
-    }
+    ///// <summary>
+    ///// Sets the type name to be used for server-side data portal
+    ///// authorization. Type must implement IAuthorizeDataPortal.
+    ///// </summary>
+    ///// <param name="typeName">Assembly qualified type name</param>
+    //public CslaDataPortalConfiguration ServerAuthorizationProviderType(string typeName)
+    //{
+    //  ConfigurationManager.AppSettings["CslaAuthorizationProvider"] = typeName;
+    //  return this;
+    //}
 
-    /// <summary>
-    /// Sets the type of interceptor invoked
-    /// by the data portal for pre- and post-processing
-    /// of each data portal invocation. Type must implement
-    /// IInterceptDataPortal.
-    /// </summary>
-    /// <param name="typeName">Assembly qualified type name</param>
-    public CslaDataPortalConfiguration InterceptorType(string typeName)
-    {
-      ConfigurationManager.AppSettings["CslaDataPortalInterceptor"] = typeName;
-      return this;
-    }
+    ///// <summary>
+    ///// Sets the type of interceptor invoked
+    ///// by the data portal for pre- and post-processing
+    ///// of each data portal invocation. Type must implement
+    ///// IInterceptDataPortal.
+    ///// </summary>
+    ///// <param name="typeName">Assembly qualified type name</param>
+    //public CslaDataPortalConfiguration InterceptorType(string typeName)
+    //{
+    //  ConfigurationManager.AppSettings["CslaDataPortalInterceptor"] = typeName;
+    //  return this;
+    //}
 
-    /// <summary>
-    /// Sets the type of interceptor invoked
-    /// by the data portal for pre- and post-processing
-    /// of each data portal invocation. Type must implement
-    /// IInterceptDataPortal.
-    /// </summary>
-    /// <param name="type">Interceptor type</param>
-    public CslaDataPortalConfiguration InterceptorType(Type type)
-    {
-      InterceptorType(type.AssemblyQualifiedName);
-      return this;
-    }
+    ///// <summary>
+    ///// Sets the type of interceptor invoked
+    ///// by the data portal for pre- and post-processing
+    ///// of each data portal invocation. Type must implement
+    ///// IInterceptDataPortal.
+    ///// </summary>
+    ///// <param name="type">Interceptor type</param>
+    //public CslaDataPortalConfiguration InterceptorType(Type type)
+    //{
+    //  InterceptorType(type.AssemblyQualifiedName);
+    //  return this;
+    //}
 
-    /// <summary>
-    /// Sets the type of the ExceptionInspector class.
-    /// Type must implement IDataPortalExceptionInspector.
-    /// </summary>
-    /// <param name="type">Inspector type</param>
-    public CslaDataPortalConfiguration ExceptionInspectorType(Type type)
-    {
-      ExceptionInspectorType(type.AssemblyQualifiedName);
-      return this;
-    }
+    ///// <summary>
+    ///// Sets the type of the ExceptionInspector class.
+    ///// Type must implement IDataPortalExceptionInspector.
+    ///// </summary>
+    ///// <param name="type">Inspector type</param>
+    //public CslaDataPortalConfiguration ExceptionInspectorType(Type type)
+    //{
+    //  ExceptionInspectorType(type.AssemblyQualifiedName);
+    //  return this;
+    //}
 
-    /// <summary>
-    /// Sets the type name of the ExceptionInspector class.
-    /// Type must implement IDataPortalExceptionInspector.
-    /// </summary>
-    /// <param name="typeName">Assembly qualified type name</param>
-    public CslaDataPortalConfiguration ExceptionInspectorType(string typeName)
-    {
-      ConfigurationManager.AppSettings["CslaDataPortalExceptionInspector"] = typeName;
-      return this;
-    }
+    ///// <summary>
+    ///// Sets the type name of the ExceptionInspector class.
+    ///// Type must implement IDataPortalExceptionInspector.
+    ///// </summary>
+    ///// <param name="typeName">Assembly qualified type name</param>
+    //public CslaDataPortalConfiguration ExceptionInspectorType(string typeName)
+    //{
+    //  ConfigurationManager.AppSettings["CslaDataPortalExceptionInspector"] = typeName;
+    //  return this;
+    //}
 
-    /// <summary>
-    /// Sets the type of the factor loader used to create
-    /// server-side instances of business object factories when using
-    /// the FactoryDataPortal model. Type must implement
-    /// IObjectFactoryLoader.
-    /// </summary>
-    /// <param name="type">Factory loader type</param>
-    public CslaDataPortalConfiguration FactoryLoaderType(Type type)
-    {
-      FactoryLoaderType(type.AssemblyQualifiedName);
-      return this;
-    }
+    ///// <summary>
+    ///// Sets the type of the factor loader used to create
+    ///// server-side instances of business object factories when using
+    ///// the FactoryDataPortal model. Type must implement
+    ///// IObjectFactoryLoader.
+    ///// </summary>
+    ///// <param name="type">Factory loader type</param>
+    //public CslaDataPortalConfiguration FactoryLoaderType(Type type)
+    //{
+    //  FactoryLoaderType(type.AssemblyQualifiedName);
+    //  return this;
+    //}
 
-    /// <summary>
-    /// Sets the type name of the factor loader used to create
-    /// server-side instances of business object factories when using
-    /// the FactoryDataPortal model. Type must implement
-    /// IObjectFactoryLoader.
-    /// </summary>
-    /// <param name="typeName">Assembly qualified type name</param>
-    public CslaDataPortalConfiguration FactoryLoaderType(string typeName)
-    {
-      ConfigurationManager.AppSettings["CslaObjectFactoryLoader"] = typeName;
-      return this;
-    }
+    ///// <summary>
+    ///// Sets the type name of the factor loader used to create
+    ///// server-side instances of business object factories when using
+    ///// the FactoryDataPortal model. Type must implement
+    ///// IObjectFactoryLoader.
+    ///// </summary>
+    ///// <param name="typeName">Assembly qualified type name</param>
+    //public CslaDataPortalConfiguration FactoryLoaderType(string typeName)
+    //{
+    //  ConfigurationManager.AppSettings["CslaObjectFactoryLoader"] = typeName;
+    //  return this;
+    //}
 
     /// <summary>
     /// Sets a value indicating whether objects should be
@@ -223,30 +218,30 @@ namespace Csla.Configuration
       return this;
     }
 
-    /// <summary>
-    /// Sets the assembly qualified type name of the dashboard, or
-    /// 'Dashboard' for default, or 'NullDashboard' for the null dashboard.
-    /// </summary>
-    /// <param name="type">Dashboard type</param>
-    /// <returns></returns>
-    public CslaDataPortalConfiguration DashboardType(Type type)
-    {
-      DashboardType(type.AssemblyQualifiedName);
-      return this;
-    }
+    ///// <summary>
+    ///// Sets the assembly qualified type name of the dashboard, or
+    ///// 'Dashboard' for default, or 'NullDashboard' for the null dashboard.
+    ///// </summary>
+    ///// <param name="type">Dashboard type</param>
+    ///// <returns></returns>
+    //public CslaDataPortalConfiguration DashboardType(Type type)
+    //{
+    //  DashboardType(type.AssemblyQualifiedName);
+    //  return this;
+    //}
 
-    /// <summary>
-    /// Sets the assembly qualified type name of the dashboard, or
-    /// 'Dashboard' for default, or 'NullDashboard' for the null dashboard.
-    /// </summary>
-    /// <param name="typeName">Assembly qualified type name</param>
-    /// <returns></returns>
-    public CslaDataPortalConfiguration DashboardType(string typeName)
-    {
-      ConfigurationManager.AppSettings["CslaDashboardType"] = typeName;
-      //Csla.Server.Dashboard.DashboardFactory.Reset();
-      return this;
-    }
+    ///// <summary>
+    ///// Sets the assembly qualified type name of the dashboard, or
+    ///// 'Dashboard' for default, or 'NullDashboard' for the null dashboard.
+    ///// </summary>
+    ///// <param name="typeName">Assembly qualified type name</param>
+    ///// <returns></returns>
+    //public CslaDataPortalConfiguration DashboardType(string typeName)
+    //{
+    //  ConfigurationManager.AppSettings["CslaDashboardType"] = typeName;
+    //  //Csla.Server.Dashboard.DashboardFactory.Reset();
+    //  return this;
+    //}
 
     /// <summary>
     /// Sets a value indicating whether any
